requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
supabase
python-dotenv
Pillow
qrcode
requests
datetime
